# backend/swagger/v1/swagger.yaml
openapi: 3.0.3
info:
  title: TeamSync API
  description: |
    TeamSync is a real-time asynchronous standup tool for remote teams.
    This API provides endpoints for authentication, team management, standup submissions, and real-time updates via WebSockets.
  version: 1.0.0
  contact:
    name: TeamSync Support
    email: support@teamsync.app
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.teamsync.app/api/v1
    description: Production server
...
paths:
...
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Retrieve the currently authenticated user's profile
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
    put:
      tags:
        - Authentication
      summary: Update profile
      description: Update the current user's profile information
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  $ref: '#/components/schemas/UpdateUserInput'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /teams:
    get:
      tags:
        - Teams
      summary: List teams
      description: Get all teams the current user is a member of
      operationId: listTeams
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of teams
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Team'
    post:
      tags:
        - Teams
      summary: Create team
      description: Create a new team with the current user as owner
      operationId: createTeam
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - team
              properties:
                team:
                  $ref: '#/components/schemas/CreateTeamInput'
      responses:
        '201':
          description: Team created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Team'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /teams/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Teams
      summary: Get team
      description: Get detailed information about a specific team
      operationId: getTeam
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Team details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Team'
        '404':
          description: Team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Teams
      summary: Update team
      description: Update team settings (admin only)
      operationId: updateTeam
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                team:
                  $ref: '#/components/schemas/UpdateTeamInput'
      responses:
        '200':
          description: Team updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Team'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Teams
      summary: Delete team
      description: Delete a team (owner only)
      operationId: deleteTeam
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Team deleted
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /teams/{slug}/invite:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    post:
      tags:
        - Teams
      summary: Invite member
      description: Invite a user to join the team by email (admin only)
      operationId: inviteMember
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [member, admin]
                  default: member
      responses:
        '201':
          description: Invitation sent
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /teams/{slug}/join:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    post:
      tags:
        - Teams
      summary: Join team
      description: Join a team using an invite code
      operationId: joinTeam
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                invite_code:
                  type: string
      responses:
        '201':
          description: Joined team successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Team'
        '422':
          description: Invalid invite code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PASETO
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        full_name:
          type: string
        initials:
          type: string
        timezone:
          type: string
        avatar_url:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, suspended]
        role:
          type: string
          enum: [member, admin, super_admin]
        created_at:
          type: string
          format: date-time
...
    UpdateUserInput:
      type: object
      properties:
        first_name:
          type: string
          maxLength: 50
        last_name:
          type: string
          maxLength: 50
        timezone:
          type: string
        avatar_url:
          type: string
...
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        timezone:
          type: string
        standup_time:
          type: string
        standup_days:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
        settings:
          type: object
        member_count:
          type: integer
        current_user_role:
          type: string
          enum: [owner, admin, member]
        today_stats:
          type: object
          properties:
            submitted:
              type: integer
            missed:
              type: integer
            total:
              type: integer
        invite_code:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        created_by:
          $ref: '#/components/schemas/User'
    CreateTeamInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
        timezone:
          type: string
        standup_time:
          type: string
        standup_days:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
    UpdateTeamInput:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        timezone:
          type: string
        standup_time:
          type: string
        standup_days:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string